// src/states/ducks/positions/types.ts
export enum PositionsActionTypes {
  GET = '@positions/GET',
  CREATE = '@position/CREATE',
  UPDATE = '@position/UPDATE',
  DELETE = '@position/DELETE',
  ADD = '@positions/ADD',
  ADD_NEW = '@position/ADD_NEW',
  EDIT = '@position/EDIT',
  ERROR = '@positions/ERROR',
  CLEAR = '@positions/CLEAR',
}